\name{rpst-internal}
\title{Internal rpst functions}
\alias{combsub.f}
\alias{df}
\alias{dif.f}
\alias{diffsplit.f}
\alias{get.class}
\alias{get.region}
\alias{new.df}
\alias{nomcomb.f}
\alias{nominal.f}
\alias{orddiffsplit.f}
\alias{replacenominal.f}
\alias{tree.f}
\alias{vartype.f}
\description{Internal rpst functions}
\usage{
rpst(data, datapath, maxlay = 12, minsize = 5)
}
\author{Yewei Li}
\details{These are not intended for use by users.}
\keyword{internal}
